Create lengthOfLongestSubstring.java
The instructions for this were: 

Given a string, find the length of the longest substring without repeating characters.

Examples:

Given "abcabcbb", the answer is "abc", which the length is 3.

Given "bbbbb", the answer is "b", with the length of 1.

Given "pwwkew", the answer is "wke", with the length of 3. Note that the answer must be a substring, "pwke" is a subsequence and not a substring

I'm working on hashmaps and using them in situations in order to reduce the complexity of my methods.